TGL-83.upd: updated shipments router for orders component on front, updated fixtures

+++ routers/shipments.ts
@@ -82,10 +82,17 @@ shipmentsRouter.post(
   },
 );
 
-shipmentsRouter.get('/', auth, permit('admin', 'super'), async (req, res) => {
+shipmentsRouter.get('/', auth, async (req: RequestWithUser, res) => {
   try {
+    const user = req.user;
     const regionId = req.query.region as string;
     const orderByTrackingNumber = req.query.orderByTrackingNumber as string;
+    const marketId = req.query.marketId as string;
+
+    if (marketId) {
+      const shipments = await Shipment.find({ userMarketId: marketId });
+      return res.send({ message: 'Список грузов одного пользователя', shipments });
+    }
 
     let filter: FilterQuery<ShipmentData> = {};
 
@@ -102,21 +109,30 @@ shipmentsRouter.get('/', auth, permit('admin', 'super'), async (req, res) => {
       return res.send(shipment);
     }
 
-    const shipments = await Shipment.find(filter).populate('userId', 'firstName lastName');
-    return res.send({ message: 'Список грузов', shipments });
+    if (user?.role === 'super' || user?.role === 'admin') {
+      const shipments = await Shipment.find(filter).populate('userId', 'firstName lastName');
+      return res.send({ message: 'Список грузов', shipments });
+    }
   } catch (e) {
     res.send(e);
   }
 });
 
-shipmentsRouter.delete('/:id', auth, permit('admin'), async (req, res, next) => {
+shipmentsRouter.delete('/:id', auth, async (req: RequestWithUser, res, next) => {
   try {
     const id = req.params.id;
+    const user = req.user;
 
     try {
       new mongoose.Types.ObjectId(id);
     } catch {
       return res.status(404).send({ error: 'Wrong ID!' });
+    }
+
+    const shipment = await Shipment.findById(id);
+
+    if (shipment?.userMarketId !== user?.marketId) {
+      return res.status(401).send({ message: 'Вы не имеете права удалять чужие грузы!' });
     }
 
     const result = await Shipment.findByIdAndDelete(id);
